fix(vulkan): hook GetPhysicalDeviceMemoryProperties2KHR alias + inflate heapUsage on budget query
Two related fixes for VK_EXT_memory_budget on Carbonite/Kit and similar
engines that target a Vulkan 1.0 instance with the KHR-extension form of
properties2:

(1) Hook the KHR alias of vkGetPhysicalDeviceMemoryProperties2

    The HAMI_HOOK macro previously matched only the core (Vulkan 1.1+)
    name. Loaders queried our layer's GetInstanceProcAddr for the KHR
    alias `vkGetPhysicalDeviceMemoryProperties2KHR` and fell through to
    the next layer / ICD when we returned NULL. Result: clamp_heaps was
    never invoked on those calls, so engines saw the host GPU's full
    heap size (~45 GiB) instead of the partition limit. Add a
    HAMI_HOOK_KHR_ALIAS macro and wire it for the memory-properties2
    entry point so the same hami_vk* function services both names.

(2) Inflate heapUsage rather than clamping heapBudget

    Engines that consume VK_EXT_memory_budget render an on-screen
    "X used / Y available" overlay computed as heapBudget - heapUsage.
    Clamping heapBudget to the partition limit produces the right
    "available" value but causes omni.physx.tensors to deadlock during
    plugin initialization (Isaac Sim Streaming 6.0, NVIDIA 580.142).
    The deadlock persists regardless of whether heapBudget is clamped
    to partition_limit, partition_limit-heapUsage, or any value below
    heap.size — PhysX/Carbonite consumes the absolute heapBudget
    through paths beyond the simple subtraction.

    Workaround: leave heapBudget at the ICD-reported value (host GPU's
    free memory) and inflate heapUsage by (icd_budget - partition_limit).
    The overlay's available calculation now equals
    heapBudget - (icd_usage + delta) = partition_limit - icd_usage,
    matching the partition. PhysX is unaffected because it does not
    consult heapUsage.

Verified end-to-end on isaac-launchable-0 (23 GiB partition,
ws-node074):

  Pre-fix:  size=44.99 GiB heapBudget=32.12 GiB
            overlay "1.3 used / 32.2 available"     [false]
  Post-fix: size=23.00 GiB heapBudget=33.4  GiB heapUsage=11.5 GiB
            overlay "11.5 used / 21.9 available"    [21.9 = 23 - 1.1]
            omni.physx.tensors initializes cleanly; Kit reaches the
            streaming-ready idle state at 60 fps.

+++ src/vulkan/budget.c
@@ -1,5 +1,6 @@
 #include "vulkan/budget.h"
 
+#include <inttypes.h>
 #include <pthread.h>
 #include <stdint.h>
 #include <stdio.h>
@@ -43,7 +44,7 @@ static void hami_core_init_once(void) { (void)cuInit(0); }
 int hami_budget_reserve(int dev, size_t size) {
     pthread_once(&g_hami_core_init, hami_core_init_once);
     uint64_t limit = hami_core_get_memory_limit(dev);
-    HAMI_TRACE("budget_reserve dev=%d size=%zu limit=%llu", dev, size, (unsigned long long)limit);
+    HAMI_TRACE("budget_reserve dev=%d size=%zu limit=%" PRIu64, dev, size, (uint64_t)limit);
     if (limit == 0) {
         /* Unlimited — skip check, but still bump the counter so metrics
          * remain accurate. add_gpu_device_memory_usage returns 0 on
@@ -68,6 +69,6 @@ void hami_budget_release(int dev, size_t size) {
 size_t hami_budget_of(int dev) {
     pthread_once(&g_hami_core_init, hami_core_init_once);
     uint64_t v = hami_core_get_memory_limit(dev);
-    HAMI_TRACE("budget_of dev=%d -> limit=%llu", dev, (unsigned long long)v);
+    HAMI_TRACE("budget_of dev=%d -> limit=%" PRIu64, dev, (uint64_t)v);
     return (size_t)v;
 }
